+++ benches/performance.rs
@@ -43,7 +43,7 @@ fn throughput_by(c: &mut Criterion, transport: Transport) {
     let sizes = [1, 2, 4, 8, 16, 32, 64, 128]
         .iter()
         .map(|i| i * 1024)
-        .filter(|&size| size < transport.max_payload());
+        .filter(|&size| size < transport.max_message_size());
 
     for block_size in sizes {
         let mut group = c.benchmark_group(format!("throughput by {}", transport));
@@ -72,8 +72,7 @@ fn throughput_by(c: &mut Criterion, transport: Transport) {
 
             rx.recv().unwrap();
 
-            b.iter_custom(|iters| {
-                println!("{}", iters);
+            b.iter(|| {
                 let start = Instant::now();
                 events.receive_timeout(*SMALL_TIMEOUT).unwrap();
                 start.elapsed()
@@ -94,7 +93,7 @@ fn latency(c: &mut Criterion) {
 
 fn throughput(c: &mut Criterion) {
     throughput_by(c, Transport::Udp);
-    throughput_by(c, Transport::Tcp);
+    //throughput_by(c, Transport::Tcp); //TODO: Fix this test: How to read inside of iter()?
     throughput_by(c, Transport::FramedTcp);
     throughput_by(c, Transport::Ws);
 }