+++ src/transport.rs
@@ -1,6 +1,6 @@
 use crate::engine::{AdapterLauncher};
 use crate::adapters::{
-    tcp::{TcpAdapter},
+    tcp::{self, TcpAdapter},
     framed_tcp::{self, FramedTcpAdapter},
     udp::{self, UdpAdapter},
     web_socket::{self, WsAdapter},
@@ -18,20 +18,20 @@ use strum::{EnumIter};
 #[derive(Clone, Copy, PartialEq, Eq, Hash, Debug)]
 pub enum Transport {
     /// The TCP stream protocol.
-    /// Note that as stream protocol, receiving a message from Tcp do not imply to read
+    /// Note that as stream protocol, receiving a message from TCP do not imply to read
     /// the entire message.
-    /// If you want a packet based way to send over tcp use [`FramedTcp`]
+    /// If you want a packet based way to send over TCP, use [`FramedTcp`]
     Tcp,
 
-    /// Like TCP, but encoded with a slim frame layer to manage the data as sized packets,
+    /// Like TCP, but encoded with a slim frame layer to manage the data as a packet,
     /// not as a stream.
     FramedTcp,
 
-    /// Udp protocol.
-    /// Note that UDP is not connection oriented, the packet can be lost or received disordered.
+    /// UDP protocol.
+    /// Note that UDP is not connection oriented, a packet can be lost or received disordered.
     Udp,
 
-    /// Websocket protocol.
+    /// WebSocket protocol.
     Ws,
 }
 
@@ -47,17 +47,20 @@ impl Transport {
         };
     }
 
-    /// Max packet payload size available for each transport
-    pub const fn max_payload(self) -> usize {
-        const UNLIMIT: usize = usize::MAX;
+    /// Max packet payload size available for each transport.
+    /// If the protocol is not packet-based (e.g. TCP, that is a stream),
+    /// the returned value correspond with the maximum bytes that can produce a read event.
+    pub const fn max_message_size(self) -> usize {
         match self {
-            Self::Tcp => UNLIMIT,
+            Self::Tcp => tcp::INPUT_BUFFER_SIZE,
             Self::FramedTcp => framed_tcp::MAX_TCP_PAYLOAD_LEN,
             Self::Udp => udp::MAX_UDP_PAYLOAD_LEN,
             Self::Ws => web_socket::MAX_WS_PAYLOAD_LEN,
         }
     }
 
+    /// Tell if the transport protocol is a connection oriented protocol.
+    /// If it is, `Connection` and `Disconnection` events will be generated.
     pub const fn is_connection_oriented(self) -> bool {
         match self {
             Transport::Tcp => true,
@@ -67,6 +70,11 @@ impl Transport {
         }
     }
 
+    /// Tell if the transport protocol is a packet based protocol.
+    /// It implies that any send call corresponds to a data message event.
+    /// The opossite of a packet based is a stream based transport (e.g Tcp).
+    /// In this case, reading a data message event do not imply reading the entire message sent.
+    /// It is in change of the user to determinate how to read the data.
     pub const fn is_packet_based(self) -> bool {
         match self {
             Transport::Tcp => false,