+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Release 0.10.0
+- Renamed `Transport::Tcp` as `Transport::FramedTcp`.
+  **WARNING**: If previously you was using `Transport::Tcp` you probably want to use now
+  `Transport::FramedTcp` (that behaves the same).
+- Added `Transport::Tcp` that has no encoding layer. Now `Transport::Tcp` is purely TCP.
+- Renamed `Transport::max_payload()` to `Transport::max_message_size()`.
+
 ## Release 0.9.4
 - Fixed ResourceId to works with 32-bits.